#134 implemented !event_create

+++ handlers/events.js
@@ -11,20 +11,23 @@ const utils = require('../utils/utils.js');
 /**
  * Create an event
  * @param {Message} msg
+ * @param {Array} msgParms
  * @param {GuildModel} guildConfig
  */
-async function handleEventCreate(msg, guildConfig) {
+async function handleEventCreate(msg, msgParms, guildConfig) {
     let eventChannelID = guildConfig.channelForEvents ? guildConfig.channelForEvents : msg.channel.id;
     try {
-        let eventString = msg.content.substring((guildConfig.prefix + 'event create').length);
-        let eventCreateResult = await bc_eventCreate(msg.member.id, eventChannelID, msg.guild.id, eventString);
+        // let eventString = msg.content.substring((guildConfig.prefix + 'event create').length);
+        let eventCreateResult = await bc_eventCreate(msg.member.id, eventChannelID, msg.guild.id, msgParms, msg);
         if (eventCreateResult) {
-            await msg.delete();
+            if (msg.deletable) {
+                await msg.delete();
+            }
         } else {
             throw new Error("Could not create event.");
         }
     } catch (error) {
-        console.error('events.handleEventCreate:', error.message);
+        console.error('handleEventCreate:', error.message);
         await utils.sendDirectOrFallbackToChannel([
             { name: 'Event Create Error', value: `${error.message}` },
             { name: 'Timezone Lookup', value: `<${Config.httpServerURL}/timezones?guildID=${msg.guild.id}&channel=${msg.channel.id}>` }
@@ -37,10 +40,11 @@ async function handleEventCreate(msg, guildConfig) {
  * @param {String} currUserId
  * @param {String} channelIDForEvent
  * @param {String} guildID
- * @param {String} eventString
+ * @param {Array} msgParms
+ * @param {Message} msg optional, pass if there was an original message for the event creation
  * @returns
  */
-async function bc_eventCreate(currUserId, channelIDForEvent, guildID, eventString) {
+async function bc_eventCreate(currUserId, channelIDForEvent, guildID, msgParms, msg) {
     try {
         let theGuild = client.guilds.cache.get(guildID);
         if (theGuild) {
@@ -48,19 +52,19 @@ async function bc_eventCreate(currUserId, channelIDForEvent, guildID, eventStrin
             if (!currUser || !currUser.timezone) {
                 throw new Error('Please set your timezone first using `timezone [YOUR TIMEZONE]`!');
             } else {
-                let eventArray = parseEventString(eventString);
-                let validatedEvent = await validateEvent(eventArray, guildID, currUser);
+                // let eventArray = parseEventString(eventString);
+                let validatedEvent = await validateEvent(msgParms, guildID, currUser);
                 let eventChannel = await theGuild.channels.resolve(channelIDForEvent);
                 let sentMessage = await eventChannel.send(await embedForEvent(theGuild.iconURL(), [validatedEvent], undefined, true));
                 validatedEvent.channelID = sentMessage.channel.id;
                 validatedEvent.messageID = sentMessage.id;
                 await validatedEvent.save();
-                await sentMessage.react('✅');
-                await sentMessage.react('❎');
-                await sentMessage.react('▶️');
-                await sentMessage.react('🕟');
-                await sentMessage.react(`\u{1F5D1}`);
-                await utils.sendDirectOrFallbackToChannel([{ name: '🗡 Event Create 🛡', value: `<@${currUserId}> - created event successfully.`, inline: true }], sentMessage, await client.users.resolve(currUserId));
+                sentMessage.react('✅');
+                sentMessage.react('❎');
+                sentMessage.react('▶️');
+                sentMessage.react('🕟');
+                sentMessage.react(`\u{1F5D1}`);
+                await utils.sendDirectOrFallbackToChannel([{ name: '🗡 Event Create 🛡', value: `<@${currUserId}> - created event successfully.`, inline: true }], msg ? msg : sentMessage, await client.users.resolve(currUserId), false, sentMessage.url);
                 return true;
             }
         } else {
@@ -325,33 +329,42 @@ async function handleEventListDeployed(msg, guildConfig) {
 
 /**
  * validate event and return object with proper types
- * @param {Array} eventArray
+ * @param {Array} msgParms
  * @param {Message} msg
  * @param {UserModel} currUser
  * @returns {EventModel}
  */
-async function validateEvent(eventArray, guildID, currUser, existingEvent) {
-    if ((!eventArray['!TITLE'] && !existingEvent && !existingEvent.title) || eventArray['!TITLE'] === null) {
+async function validateEvent(msgParms, guildID, currUser, existingEvent) {
+    let etitle = msgParms.find(p => p.name == 'title') ? msgParms.find(p => p.name == 'title').value : undefined;
+    let efor = msgParms.find(p => p.name == 'for') ? msgParms.find(p => p.name == 'for').value : undefined;
+    let eon = msgParms.find(p => p.name == 'on') ? msgParms.find(p => p.name == 'on').value : undefined;
+    let eat = msgParms.find(p => p.name == 'at') ? msgParms.find(p => p.name == 'at').value : undefined;
+    let ewith = msgParms.find(p => p.name == 'with') ? msgParms.find(p => p.name == 'with').value : undefined;
+    let edesc = msgParms.find(p => p.name == 'desc') ? msgParms.find(p => p.name == 'desc').value : undefined;
+    let edmgm = msgParms.find(p => p.name == 'dmgm') ? msgParms.find(p => p.name == 'dmgm').value : undefined;
+    let ecampaign = msgParms.find(p => p.name == 'campaign') ? msgParms.find(p => p.name == 'campaign').value : undefined;
+
+    if ((!etitle && !existingEvent && !existingEvent.title) || etitle === null) {
         throw new Error('You must include a title for your event.');
-    } else if ((!eventArray['!FOR'] && eventArray['!FOR'] === null && !existingEvent && !existingEvent.duration) || eventArray['!FOR'] === null) {
+    } else if ((!efor && (!existingEvent || (existingEvent && !existingEvent.duration))) || efor === null) {
         throw new Error('You must include a duration for your event.');
-    } else if (eventArray['!FOR'] && isNaN(eventArray['!FOR'])) {
-        throw new Error(`The duration hours needs to be a number, not: "${eventArray['!FOR']}"`);
-    } else if ((!eventArray['!ON'] && eventArray['!ON'] === null && !existingEvent && !existingEvent.date_time) || eventArray['!ON'] === null) {
+    } else if (efor && isNaN(efor)) {
+        throw new Error(`The duration hours needs to be a number, not: "${efor}"`);
+    } else if ((!eon && (!existingEvent || (existingEvent && !existingEvent.date_time))) || eon === null) {
         throw new Error('You must include a date for your event.');
-    } else if ((!eventArray['!AT'] && eventArray['!AT'] === null && !existingEvent && !existingEvent.date_time) || eventArray['!AT'] === null) {
+    } else if ((!eat && (!existingEvent || (existingEvent && !existingEvent.date_time))) || eat === null) {
         throw new Error('You must include a time for your event.');
-    } else if ((!eventArray['!WITH'] && eventArray['!WITH'] === null && !existingEvent && !existingEvent.number_player_slots) || eventArray['!WITH'] === null) {
+    } else if ((!ewith && (!existingEvent || (existingEvent && !existingEvent.number_player_slots))) || ewith === null) {
         throw new Error('You must include a number of player slots for your event.');
-    } else if ((!eventArray['!DESC'] && eventArray['!DESC'] === null && !existingEvent && !existingEvent.description) || eventArray['!DESC'] === null) {
+    } else if ((!edesc && (!existingEvent || (existingEvent && !existingEvent.description))) || edesc === null) {
         throw new Error('You must include a description for your event.');
-    } else if (eventArray['!WITH'] && isNaN(eventArray['!WITH'])) {
-        throw new Error(`The number of player slots needs to be a number, not: "${eventArray['!WITH']}"`);
+    } else if (ewith && isNaN(ewith)) {
+        throw new Error(`The number of player slots needs to be a number, not: "${ewith}"`);
     }
 
     let validatedEvent = existingEvent ? existingEvent : new EventModel({ guildID: guildID, userID: currUser.userID });
 
-    if (eventArray['!ON'] || eventArray['!AT']) {
+    if (eon || eat) {
         let timezoneOffset = getTimeZoneOffset(currUser.timezone);
         console.log('tz offset: ' + timezoneOffset);
 
@@ -363,8 +376,8 @@ async function validateEvent(eventArray, guildID, currUser, existingEvent) {
             // console.log('usersoriginaleventdate %s', usersOriginalEventDate);
         }
         //@todo: figure out a way to get the fuzzy parsing per date and per time working
-        let onDate = eventArray['!ON'] ? eventArray['!ON'] : formatJustDate(usersOriginalEventDate);
-        let atTime = eventArray['!AT'] ? eventArray['!AT'] : formatJustTime(usersOriginalEventDate);
+        let onDate = eon ? eon : formatJustDate(usersOriginalEventDate);
+        let atTime = eat ? eat : formatJustTime(usersOriginalEventDate);
         let dateTimeStringToParse = `${onDate} at ${atTime}`;
         let refDate = usersOriginalEventDate ? usersOriginalEventDate : getDateInDifferentTimezone(new Date(), currUser.timezone);//new Date(new Date().toLocaleString("en-US", { timeZone: currUser.timezone }));
         // console.log('refDate %s then - on %s at %s', refDate, onDate, atTime);
@@ -374,18 +387,18 @@ async function validateEvent(eventArray, guildID, currUser, existingEvent) {
     }
 
     // console.log(eventArray);
-    validatedEvent.title = eventArray['!TITLE'] === null ? undefined : (eventArray['!TITLE'] ? eventArray['!TITLE'] : validatedEvent.title);
-    validatedEvent.dm = eventArray['!DMGM'] === null ? undefined : (eventArray['!DMGM'] ? eventArray['!DMGM'] : validatedEvent.dm);
-    validatedEvent.duration_hours = eventArray['!FOR'] === null ? undefined : (eventArray['!FOR'] ? eventArray['!FOR'] : validatedEvent.duration_hours);
-    validatedEvent.number_player_slots = eventArray['!WITH'] === null ? undefined : (eventArray['!WITH'] ? eventArray['!WITH'] : validatedEvent.number_player_slots);
-    validatedEvent.campaign = eventArray['!CAMPAIGN'] === null ? undefined : (eventArray['!CAMPAIGN'] ? eventArray['!CAMPAIGN'] : validatedEvent.campaign);
-    validatedEvent.description = eventArray['!DESC'] === null ? undefined : (eventArray['!DESC'] ? eventArray['!DESC'] : validatedEvent.description);
-    let changeDetected = false;
-    for (let key of Object.keys(eventArray)) {
-        if (eventArray[key]) {
-            changeDetected = true;
-        }
-    }
+    validatedEvent.title = etitle === null ? undefined : (etitle ? etitle : validatedEvent.title);
+    validatedEvent.dm = edmgm === null ? undefined : (edmgm ? edmgm : validatedEvent.dm);
+    validatedEvent.duration_hours = efor === null ? undefined : (efor ? efor : validatedEvent.duration_hours);
+    validatedEvent.number_player_slots = ewith === null ? undefined : (ewith ? ewith : validatedEvent.number_player_slots);
+    validatedEvent.campaign = ecampaign === null ? undefined : (ecampaign ? ecampaign : validatedEvent.campaign);
+    validatedEvent.description = edesc === null ? undefined : (edesc ? edesc : validatedEvent.description);
+    let changeDetected = msgParms.length > 0;
+    // for (let key of Object.keys(eventArray)) {
+    //     if (eventArray[key]) {
+    //         changeDetected = true;
+    //     }
+    // }
     if (!changeDetected) {
         throw new Error("No changes where detected, please verify your `event edit` command");
     }
@@ -498,7 +511,7 @@ async function embedForEvent(guildIconURL, eventArray, title, isShow) {
                 .setColor(utils.COLORS.BLUE);
             i = 0;
         }
-        let dmgmString = theEvent.dm ? theEvent.dm : 'Unassigned';
+        let dmgmString = theEvent.dm ? '<@' + theEvent.dm + '>' : 'Unassigned';
         let messageTitleAndUrl = isShow
             ? `${theEvent.title} id: ${theEvent._id}`
             : `${getEmbedLinkForEvent(theEvent)}`;
